+++ Cargo.toml
@@ -25,12 +25,13 @@ vhost_user_block = { path = "vhost_user_block"}
 vhost_user_fs = { path = "vhost_user_fs"}
 vhost_user_net = { path = "vhost_user_net"}
 virtio-bindings = { version = "0.1", features = ["virtio-v5_0_0"]}
+virtio-devices = { path = "virtio-devices" }
+vhost_rs = { git = "https://github.com/cloud-hypervisor/vhost", branch = "dragonball", package = "vhost", features = ["vhost-user-slave"] }
 vmm = { path = "vmm" }
+vmm-sys-util = "0.6.1"
 vm-device = { path = "vm-device" }
 vm-memory = "0.2.1"
-vmm-sys-util = "0.6.1"
 vm-virtio = { path = "vm-virtio" }
-vhost_rs = { git = "https://github.com/cloud-hypervisor/vhost", branch = "dragonball", package = "vhost", features = ["vhost-user-slave"] }
 
 [patch.crates-io]
 vm-memory = { git = "https://github.com/cloud-hypervisor/vm-memory", branch = "ch" }
@@ -62,6 +63,7 @@ members = [
     "devices",
     "qcow",
     "pci",
+    "virtio-devices",
     "vmm",
     "vm-virtio",
     "vm-device",
