+++ qcow/Cargo.toml
@@ -14,7 +14,7 @@ libc = "0.2.71"
 log = "0.4.8"
 remain = "0.2.2"
 vmm-sys-util = ">=0.3.1"
-vm-virtio = { path = "../vm-virtio" }
+virtio-devices = { path = "../virtio-devices" }
 
 [dev-dependencies]
 tempfile = "3.1.0"
