+++ vhost_user_net/Cargo.toml
@@ -13,7 +13,7 @@ net_util = { path = "../net_util" }
 vhost_user_backend = { path = "../vhost_user_backend" }
 vhost_rs = { git = "https://github.com/cloud-hypervisor/vhost", branch = "dragonball", package = "vhost", features = ["vhost-user-slave"] }
 virtio-bindings = "0.1.0"
+virtio-devices = { path = "../virtio-devices" }
 vm-memory = "0.2.1"
-vm-virtio = { path = "../vm-virtio" }
 vmm = { path = "../vmm" }
 vmm-sys-util = ">=0.3.1"
