+++ virtio-devices/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "virtio-devices"
+version = "0.1.0"
+authors = ["The Cloud Hypervisor Authors"]
+edition = "2018"
+
+[features]
+default = []
+pci_support = ["pci"]
+mmio_support = []
+
+[dependencies]
+anyhow = "1.0"
+arc-swap = ">=0.4.4"
+byteorder = "1.3.4"
+devices = { path = "../devices" }
+epoll = ">=4.0.1"
+libc = "0.2.71"
+log = "0.4.8"
+net_gen = { path = "../net_gen" }
+net_util = { path = "../net_util" }
+pci = { path = "../pci", optional = true }
+serde = ">=1.0.27"
+serde_derive = ">=1.0.27"
+serde_json = ">=1.0.9"
+tempfile = "3.1.0"
+vfio-ioctls = { git = "https://github.com/cloud-hypervisor/vfio-ioctls", branch = "ch" }
+vhost_rs = { git = "https://github.com/cloud-hypervisor/vhost", branch = "dragonball", package = "vhost", features = ["vhost-user-master", "vhost-user-slave"] }
+virtio-bindings = { version = "0.1", features = ["virtio-v5_0_0"]}
+vm-allocator = { path = "../vm-allocator" }
+vm-device = { path = "../vm-device" }
+vm-memory = { version = "0.2.1", features = ["backend-mmap", "backend-atomic"] }
+vm-migration = { path = "../vm-migration" }
+vm-virtio = { path = "../vm-virtio" }
+vmm-sys-util = ">=0.3.1"
+
