+++ virtio-devices/src/net_util.rs
