+++ vmm/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2018"
 [features]
 default = []
 acpi = ["acpi_tables","devices/acpi", "arch/acpi"]
-pci_support = ["pci", "vfio-ioctls", "vm-virtio/pci_support"]
-mmio_support = ["vm-virtio/mmio_support"]
+pci_support = ["pci", "vfio-ioctls", "virtio-devices/pci_support"]
+mmio_support = ["virtio-devices/mmio_support"]
 cmos = ["devices/cmos"]
 fwdebug = ["devices/fwdebug"]
 
@@ -34,6 +34,7 @@ serde_derive = ">=1.0.27"
 serde_json = ">=1.0.9"
 url = "2.1.1"
 vfio-ioctls = { git = "https://github.com/cloud-hypervisor/vfio-ioctls", branch = "ch", optional = true }
+virtio-devices = {path = "../virtio-devices"}
 vm-allocator = { path = "../vm-allocator" }
 vm-device = { path = "../vm-device" }
 vm-memory = { version = "0.2.1", features = ["backend-mmap", "backend-atomic"] }
